+++ packages/react-headless/tabs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @seed-design/react-tabs
 
+## 0.0.0-alpha-20241203133940
+
+### Patch Changes
+
+- tabs 자잘한 스타일 수정 및 fixTriggerList prop 추가
+
 ## 0.0.0-alpha-20241202094027
 
 ### Patch Changes
@@ -10,7 +16,7 @@
 
 ### Patch Changes
 
-- add --seed-design-current-tab-enabled-index css variable in root
+- add --seed-design-tabs-current-tab-enabled-index css variable in root
 
 ## 0.0.0-alpha-20241202030714